Numpy deprecation int float bool (#2464)
* replace numpy.int with int because of the following deprecation warning:

DeprecationWarning: `np.int` is a deprecated alias for the builtin `int`. To silence this warning, use `int` by itself. Doing this will not modify any behavior and is safe. When replacing `np.int`, you may wish to use e.g. `np.int64` or `np.int32` to specify the precision. If you wish to review your current use, check the release note link for additional information.
Deprecated in NumPy 1.20; for more details and guidance: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
     from numpy import (exp, inf, pi, sqrt, floor, sin, cos, around, int,

* Replace numpy.float with float because of the following deprecation warning:

DeprecationWarning: `np.float` is a deprecated alias for the builtin `float`. To silence this warning, use `float` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.float64` here.
 Deprecated in NumPy 1.20; for more details and guidance: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
dcalib = numpy.array([1, 1.3, 2, 3.5, 4, 5, 1.3, 6, 9.1], dtype=numpy.float)

* Replaced numpy.bool with bool because of the following depracation warning:

DeprecationWarning: `np.bool` is a deprecated alias for the builtin `bool`. To silence this warning, use `bool` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.bool_` here.
Deprecated in NumPy 1.20; for more details and guidance: https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
elif arr.dtype in np.sctypes['uint'] + [np.bool]:

+++ src/odemis/acq/stream/_projection.py
@@ -1429,7 +1429,7 @@ class LineSpectrumProjection(RGBProjection):
             # original data were pick on each line. Currently if some pixels fall
             # out of the original data, the outside pixels count as 0.
             # force the intermediate values to float, as mean() still needs to run
-            spec1d_w = ndimage.map_coordinates(spec2d, coord, output=numpy.float, order=1)
+            spec1d_w = ndimage.map_coordinates(spec2d, coord, output=float, order=1)
             spec1d = spec1d_w.mean(axis=0)
         assert spec1d.shape == (n, spec2d.shape[0])
